+++ src/ships/enemy.rs
@@ -21,6 +21,53 @@ pub struct Targeting {
     pub angle: f32,
 }
 
+pub struct DamagedBy {
+    pub attacker: Entity,
+    pub damage: u32,
+}
+#[derive(Component)]
+pub struct Adversaries(pub Vec<DamagedBy>);
+
+// NOTE: We might add a way for the enemy to be healed (healer class drone?)
+//       Since the intention is to use the Adversaries damage to determine
+//       what percentage of the damage done was by the player... we would
+//       want to rethink this to avoid a single enemy being bled for inf%
+//       XP more than they ought to be.
+//
+//       Since we also want the enemy to remember who their greatest adversary
+//       has been however, we don't want to delete damage done as it is healed...
+//
+//       So we could simply cap the damage-based XP reward at 100%. This seems fair.
+//       If you, the player, have done 10_000 damage, for example... you should be
+//       rewarded for that just the same regardless of whether the enemy has healed
+//       and some other adversary has done 30_000... no?
+//
+//       An alternative approach would be to base the XP on the relative proportionality
+//       of damages done. That is, a percentage of the total damage delt rather than of
+//       total health (which could differ). But this could result in very little reward
+//       for the same amount of work... the other approach seems fairer.
+//
+//       Imagine if it were a multiplayer scenario. You could technically have both
+//       players earn maximum HP by allowing the enemy to heal before finishing it.
+//       Does that seem like the right way to go about it? I think so.
+//
+//       Another scenario: The player has delt some damage, the enemy has gone away and
+//       been fully healed and returns as if it was never damaged in the first place...
+//       Should this be treated as a new encounter? Or should the previous damage remain?
+//       What if it has been X amount of time since you last encountered this enemy? When
+//       the game is an RPG with named characters, this could be weeks, months, years...
+//       The enemy _should_ remember that you, the player, were responsible for damage
+//       done... but the XP potential, perhaps that should be forgotten.
+//
+//       Propose we also have a concept of "Current Conflict". While in this mode, it
+//       all counts! When the conflict is over - due to fleeing or whatever - the XP
+//       potential can be forgotten. But the memory that you both fought should remain
+//       (particularly when the game is a more complete RPG).
+//
+//       It's sort of like a short-term memory and long-term memory. Write the short-term
+//       current conflict values first, then record these in the long-term memory for
+//       retrieval/use at a later time.
+
 /// The setup function
 pub(crate) fn spawn_enemies(mut commands: Commands, sprites: Res<SpriteAssets>) {
     // Spawns enemy ships
@@ -64,6 +111,7 @@ pub(crate) fn spawn_enemies(mut commands: Commands, sprites: Res<SpriteAssets>) 
                 angular_damping: 2.0,
             },
             Name::new("Enemy"),
+            Adversaries(Vec::new()),
         ));
     }
 }