+++ src/ships/ship.rs
@@ -1,9 +1,18 @@
 use bevy::prelude::*;
 use bevy_rapier2d::prelude::*;
 
-use crate::systems::{events::BulletShipContactEvent, states::GameState};
+use crate::{
+    core::resources::score::Score,
+    systems::{
+        events::{BulletShipContactEvent, DeathEvent},
+        states::GameState,
+    },
+};
 
-use super::player::Player;
+use super::{
+    enemy::{Adversaries, DamagedBy},
+    player::Player,
+};
 
 /// Ship component
 #[derive(Component)]
@@ -47,16 +56,79 @@ pub(crate) fn bullet_timers_system(time: Res<Time>, mut ship: Query<&mut Ship>) 
 pub(crate) fn ship_damage(
     mut commands: Commands,
     mut bullet_ship_contact_events: EventReader<BulletShipContactEvent>,
-    mut ship_health: Query<&mut Health, With<Ship>>,
+    mut ship: Query<&mut Health, With<Ship>>,
+    mut score: ResMut<Score>,
+    player: Query<Entity, (With<Player>, With<Ship>)>,
+    mut death_events: EventWriter<DeathEvent>,
 ) {
+    let Ok(player) = player.get_single() else {
+        return;
+    };
+
     for event in bullet_ship_contact_events.read() {
         commands.entity(event.bullet).despawn();
 
-        if let Ok(mut health) = ship_health.get_mut(event.ship) {
+        if let Ok(mut health) = ship.get_mut(event.ship) {
+            if event.bullet_spawner == player {
+                score.0 += 100;
+            }
+
             health.0 -= 100.0;
             if health.0 <= 0. {
-                commands.entity(event.ship).despawn();
+                death_events.send(DeathEvent { entity: event.ship });
             }
+        }
+    }
+}
+
+pub(crate) fn ship_death_handling(
+    mut commands: Commands,
+    mut death_events: EventReader<DeathEvent>,
+    player: Query<Entity, (With<Player>, With<Ship>)>,
+    mut score: ResMut<Score>,
+    adversaries: Query<&Adversaries, With<Ship>>,
+) {
+    let Ok(player) = player.get_single() else {
+        return;
+    };
+
+    for event in death_events.read() {
+        let Ok(adversaries) = adversaries.get(event.entity) else {
+            return;
+        };
+
+        let player_dealt_damage: u32 = adversaries
+            .0
+            .iter()
+            .filter_map(|a| {
+                if a.attacker == player {
+                    Some(a.damage)
+                } else {
+                    None
+                }
+            })
+            .sum();
+
+        // If the destroyed ship is not the player, award XP to the player equal to
+        // 10x the damage the player has dealt to the ship.
+        if event.entity != player {
+            score.0 += 10 * player_dealt_damage; // TODO: Make proportionate to player-dealt damage relative to MaxHealth
+        }
+
+        commands.entity(event.entity).despawn();
+    }
+}
+
+pub(crate) fn adversary_system(
+    mut bullet_ship_contact_events: EventReader<BulletShipContactEvent>,
+    mut ship: Query<&mut Adversaries, With<Ship>>,
+) {
+    for event in bullet_ship_contact_events.read() {
+        if let Ok(mut adversaries) = ship.get_mut(event.ship) {
+            adversaries.0.push(DamagedBy {
+                attacker: event.bullet_spawner,
+                damage: 100,
+            });
         }
     }
 }