+++ src/systems/mod.rs
@@ -186,7 +186,8 @@ impl Plugin for SystemsPlugin {
                 hud::speedometer::hud_speedometer,
                 hud::health::hud_health,
                 hud::nav::current_location,
-                hud::time::current_time,
+                // hud::time::current_time,
+                hud::score::current_score,
                 // TODO: This clumsily resolves a crash to desktop where enemy_targeting_system
                 //       was happening after ship_damage some of the time.
                 //       The SystemSets aren't aware of one another, they need to be configured.
@@ -196,7 +197,9 @@ impl Plugin for SystemsPlugin {
                 bullet::spawn_bullet.after(MovementSet),
                 (enemy::enemy_weapons_system, events::contact_system).in_set(AttackSet),
                 (
+                    ship::adversary_system.before(ship::ship_damage),
                     ship::ship_damage,
+                    ship::ship_death_handling.after(ship::ship_damage),
                     damage::ui_spawn_damage,
                     damage::ui_text_fade_out,
                 )